data vis update
still have to work on data vis stuff tmrw morning but this is what i did tonight. began working on the data viz stuff proper, imported the csv file and started working on the "bleaching" box effect. its commented out rn. tomorrow i'm going to work on adding a timer and maybe randomizing how it appears visually on the screen? we'll see if i can figure it out :D yipee!